+++ src/app/homepage/homepage-scenario-a.demo.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector:    'soho-homepage-scenario-a-demo',
+  templateUrl: './homepage-scenario-a.demo.html',
+})
+export class HomePageScenarioADemoComponent {
+}