+++ src/app/homepage/homepage-scenario-c.demo.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector:    'soho-homepage-scenario-c-demo',
+  templateUrl: './homepage-scenario-c.demo.html',
+})
+export class HomePageScenarioCDemoComponent {
+}