+++ src/soho/homepage/index.ts
@@ -1,0 +1,6 @@
+export * from './soho-homepage.component';
+export * from './soho-widget.component';
+export * from './soho-widget-content.component';
+export * from './soho-widget-header.component';
+export * from './soho-widget-title.component';
+export * from './soho-homepage.module';