+++ src/soho/homepage/soho-homepage.module.ts
@@ -1,0 +1,28 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import {
+  SohoHomePageComponent,
+  SohoWidgetComponent,
+  SohoWidgetContentComponent,
+  SohoWidgetHeaderComponent,
+  SohoWidgetTitleComponent } from './';
+
+@NgModule({
+  imports:      [ CommonModule ],
+  declarations: [
+    SohoHomePageComponent,
+    SohoWidgetComponent,
+    SohoWidgetContentComponent,
+    SohoWidgetHeaderComponent,
+    SohoWidgetTitleComponent
+  ],
+  exports:      [
+    SohoHomePageComponent,
+    SohoWidgetComponent,
+    SohoWidgetContentComponent,
+    SohoWidgetHeaderComponent,
+    SohoWidgetTitleComponent
+  ]
+})
+export class SohoHomePageModule {
+}