+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 
 ## Whats New
+* 2017-07-28 - BTH - Improvded Veto Support for modal and message dialogs.
+  - Added `dialogRef` as an argument to `beforeClose` to provide access to dialog properties.
 * 2017-07-23 - BTH - Added `dirtyRows` to dataGrid.
   - Removed `getDirtyRows(...)` as this was not implemented and did not match the underlying control.
+* 2017-07-05 - MH - Home Page (https://jira/browse/SOHO-6468)
+  - Added basic version of the Home Page
 * 2017-06-12 - PWP - Updates to `soho-expandablearea.component` and `soho-button.component`
   - Added `soho-expandable-footer` element
   - Changed `soho-expandable-header` to be optional
@@ -51,6 +55,10 @@
  * SohoListViewComponent now uses Soho Control defaults for 'options'
 
 ### Breaking Changes
+
+* 2017-07-28 - BTH - Added generic type argument onto `SohoModalDialogVetoableEventGuard`.  
+ 
+    Any code using this interface should add the generic type of the dialog component, or `any` for messages.
 
 * 2017-07-23 - BTH - Replaced `getDirtyRows(...)` with `dirtyRows()` on datagrid.